feat!: rewrite role system, fix many bugs, improved code, change migration

+++ database/migrations/2019_03_27_184418_create_equipment_file_table.php
@@ -14,8 +14,8 @@ class CreateEquipmentFileTable extends Migration
     public function up()
     {
         Schema::create('equipment_file', function (Blueprint $table) {
-            $table->unsignedInteger('equipment_id');
-            $table->unsignedInteger('file_id');
+            $table->unsignedBigInteger('equipment_id');
+            $table->unsignedBigInteger('file_id');
 
             $table->foreign('equipment_id')->references('id')->on('equipments')
                 ->onDelete('cascade');
